requirements.txt
dnspython==1.16.0
Pillow==7.1.2
pytesseract==0.3.4
selenium==3.141.0
urllib3==1.25.9
